Chatbot_RAG/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
evaluate==0.4.5
faiss-cpu==1.9.0.post1
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.12.0
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
idna==3.10
jinja2==3.1.5
json_repair==0.48.0
joblib==1.4.2
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.4
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pyyaml==6.0.2
ragas==0.3.1
regex==2024.11.6
requests==2.32.3
streamlit==1.48.1
safetensors==0.5.0
scikit-learn==1.6.0
scipy==1.15.0
sentence-transformers==3.3.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.4.1
tqdm==4.67.1
transformers==4.47.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
